+++ CHANGES.rst
@@ -23,7 +23,11 @@ Changelog of threedi-modelchecker
   argument. The check output can also be dumped to a file using ``--file``.
 
 - Compatibility fix with rasterio 1.3.6.
+
 - Drop SQLAlchemy 1.3 support, add 2.0 support.
+
+- Add check 326: this gives an info message if a record exists in the simple_infiltration
+  table, but is not referenced from the global settings.
 
 
 1.0.1 (2023-02-02)